Web nav mode is now available.
inspired by Enhanced Touch Gestures: a new mode for navigating web elements (browse mode for the most part) is now available. Due to limitations with input core, this will not be the last mode (scan is last). With this mode, one can press up or down arrow to change nav element and left or right arrows to navigate between elements. If one is not using browse mode, touch in flow command is used.